테스트
날씨 통신 테스트 및 카카오톡 연동 추가

+++ app/src/main/java/com/jiyoung/rxandroidexamples/net/model/Wind.java
@@ -7,7 +7,7 @@ package com.jiyoung.rxandroidexamples.net.model;
 public class Wind {
 
     private float speed;
-    private int deg;
+    private float deg;
 
     public float getSpeed() {
         return speed;
@@ -17,7 +17,7 @@ public class Wind {
         this.speed = speed;
     }
 
-    public int getDeg() {
+    public float getDeg() {
         return deg;
     }
